update: Refactored structure of the code.
- Refactored structure of the code.
