+++ src/datagathering/DataGather.java
@@ -9,20 +9,28 @@ import org.jsoup.nodes.Document;
 import org.jsoup.select.Elements;
 
 public class DataGather {
-	
+	public int error = 0;
 	public Document getDataInXML() {
 		Document doc = null; 
 		try {
 			doc = Jsoup.connect("http://rates.fxcm.com/RatesXML").get();
 		} catch (IOException e) {
-			e.printStackTrace();
+			doc = null;
 		}
 		return doc;
 	}
 	
 	public Map<String,Symbol> ParseSymbol(Document doc) {
-		Elements nodeList =  doc.getElementsByTag("Rate");
-		Map<String,Symbol> symbols = new HashMap<String,Symbol>();
+		Elements nodeList;
+		Map<String,Symbol> symbols;
+		
+		try {
+			nodeList =  doc.getElementsByTag("Rate");
+			symbols = new HashMap<String,Symbol>();
+		} catch (Exception e) {
+			this.error = -1;
+			return null;
+		}
 		
 		for (int i = 0; i < nodeList.size(); i++) {
 			String name = nodeList.get(i).getAllElements().attr("symbol").toString();
